backend/pyproject.toml
[project]
name = "reasoning-honesty-lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.2",
    "tinker>=0.19.0",
    "tinker-cookbook>=0.4.0",
    "uvicorn>=0.46.0",
]
